requirements_dev.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile --output-file=requirements_dev.txt requirements_dev.in
#
asgiref==3.4.1
    # via django
bleach==4.1.0
    # via readme-renderer
bump2version==1.0.1
    # via bumpversion
bumpversion==0.6.0
    # via -r requirements_dev.in
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.7
    # via requests
colorama==0.4.4
    # via
    #   twine
    #   zest.releaser
django==3.2.9
    # via
    #   -r requirements_dev.in
    #   model-bakery
docutils==0.18
    # via readme-renderer
factory-boy==3.2.1
    # via -r requirements_dev.in
faker==9.8.0
    # via
    #   -r requirements_dev.in
    #   factory-boy
freezegun==1.1.0
    # via -r requirements_dev.in
idna==3.3
    # via requests
importlib-metadata==4.8.1
    # via
    #   keyring
    #   twine
invoke==1.6.0
    # via -r requirements_dev.in
keyring==23.2.1
    # via twine
model-bakery==1.3.3
    # via -r requirements_dev.in
packaging==21.2
    # via bleach
pkginfo==1.7.1
    # via twine
pygments==2.10.0
    # via readme-renderer
pyparsing==2.4.7
    # via packaging
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
pytz==2021.3
    # via django
readme-renderer==30.0
    # via twine
requests==2.26.0
    # via
    #   requests-toolbelt
    #   twine
    #   zest.releaser
requests-toolbelt==0.9.1
    # via twine
rfc3986==1.5.0
    # via twine
six==1.16.0
    # via
    #   bleach
    #   python-dateutil
    #   zest.releaser
sqlparse==0.4.2
    # via django
text-unidecode==1.3
    # via faker
tqdm==4.62.3
    # via twine
twine==3.5.0
    # via
    #   -r requirements_dev.in
    #   zest.releaser
typing-extensions==3.10.0.2
    # via
    #   asgiref
    #   faker
    #   importlib-metadata
urllib3==1.26.7
    # via requests
webencodings==0.5.1
    # via bleach
wheel==0.37.0
    # via -r requirements_dev.in
zest.releaser==6.22.2
    # via -r requirements_dev.in
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
